# exports/al-yamen-mysql-deployment/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary User login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  role: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  role: zod.string(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  annualRevenue: zod.number(),
  totalTransactions: zod.number(),
  growthRate: zod.number(),
  activeStaff: zod.number(),
  onLeave: zod.number(),
  recentTransactions: zod.array(
    zod.object({
      id: zod.number(),
      date: zod.string(),
      type: zod.string(),
      customer: zod.string(),
      description: zod.string(),
      amount: zod.number(),
      createdAt: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Get monthly revenue data for chart
 */
export const GetRevenueChartResponseItem = zod.object({
  month: zod.string(),
  revenue: zod.number(),
});
export const GetRevenueChartResponse = zod.array(GetRevenueChartResponseItem);

/**
 * @summary List all transactions
 */
export const listTransactionsQueryLimitDefault = 100;

export const ListTransactionsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listTransactionsQueryLimitDefault),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  type: zod.string(),
  customer: zod.string(),
  description: zod.string(),
  amount: zod.number(),
  createdAt: zod.string().optional(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a new transaction
 */
export const CreateTransactionBody = zod.object({
  date: zod.string(),
  type: zod.string(),
  customer: zod.string(),
  description: zod.string(),
  amount: zod.number(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all orders
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  customer: zod.string(),
  date: zod.string(),
  total: zod.number(),
  status: zod.string(),
  createdAt: zod.string().optional(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  customer: zod.string(),
  date: zod.string(),
  total: zod.number(),
  status: zod.string().optional(),
});

/**
 * @summary List all customers
 */
export const ListCustomersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  company: zod.string(),
  phone: zod.string(),
  lastOrder: zod.string().optional(),
  totalSpent: zod.number(),
});
export const ListCustomersResponse = zod.array(ListCustomersResponseItem);

/**
 * @summary Create a new customer
 */
export const CreateCustomerBody = zod.object({
  name: zod.string(),
  company: zod.string(),
  phone: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  stock: zod.number(),
  price: zod.number(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Add a new product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  stock: zod.number(),
  price: zod.number(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List product categories with counts
 */
export const ListCategoriesResponseItem = zod.object({
  name: zod.string(),
  count: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List all employees
 */
export const ListEmployeesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
  status: zod.string(),
  salary: zod.number(),
  department: zod.string().optional(),
});
export const ListEmployeesResponse = zod.array(ListEmployeesResponseItem);

/**
 * @summary Add a new employee
 */
export const CreateEmployeeBody = zod.object({
  name: zod.string(),
  role: zod.string(),
  status: zod.string().optional(),
  salary: zod.number(),
  department: zod.string().optional(),
});

/**
 * @summary Get daily ledger balance
 */
export const GetDailyLedgerResponse = zod.object({
  totalBalance: zod.number(),
  totalIncome: zod.number(),
  totalExpenses: zod.number(),
  date: zod.string(),
});

/**
 * @summary Get expense report data
 */
export const GetExpenseReportResponseItem = zod.object({
  category: zod.string(),
  amount: zod.number(),
});
export const GetExpenseReportResponse = zod.array(GetExpenseReportResponseItem);

/**
 * @summary Get payroll breakdown
 */
export const GetPayrollDataResponseItem = zod.object({
  category: zod.string(),
  amount: zod.number(),
});
export const GetPayrollDataResponse = zod.array(GetPayrollDataResponseItem);