# exports/al-yamen-mysql-deployment/lib/api-zod/src/generated/types/authuser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  username: string;
  name: string;
  role: string;
}