# exports/al-yamen-mysql-deployment/lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Transaction {
  id: number;
  date: string;
  type: string;
  customer: string;
  description: string;
  amount: number;
  createdAt?: string;
}