# exports/al-yamen-ready-to-host/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  name: string;
  role: string;
}

export interface Transaction {
  id: number;
  date: string;
  type: string;
  customer: string;
  description: string;
  amount: number;
  createdAt?: string;
}

export interface DashboardSummary {
  annualRevenue: number;
  totalTransactions: number;
  growthRate: number;
  activeStaff: number;
  onLeave: number;
  recentTransactions: Transaction[];
}

export interface RevenuePoint {
  month: string;
  revenue: number;
}

export interface CreateTransactionBody {
  date: string;
  type: string;
  customer: string;
  description: string;
  amount: number;
}

export interface Order {
  id: number;
  orderId: string;
  customer: string;
  date: string;
  total: number;
  status: string;
  createdAt?: string;
}

export interface CreateOrderBody {
  customer: string;
  date: string;
  total: number;
  status?: string;
}

export interface Customer {
  id: number;
  name: string;
  company: string;
  phone: string;
  lastOrder?: string;
  totalSpent: number;
}

export interface CreateCustomerBody {
  name: string;
  company: string;
  phone: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  stock: number;
  price: number;
}

export interface CreateProductBody {
  name: string;
  category: string;
  stock: number;
  price: number;
}

export interface Category {
  name: string;
  count: number;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  status: string;
  salary: number;
  department?: string;
}

export interface CreateEmployeeBody {
  name: string;
  role: string;
  status?: string;
  salary: number;
  department?: string;
}

export interface LedgerSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  date: string;
}

export interface ExpenseItem {
  category: string;
  amount: number;
}

export interface PayrollItem {
  category: string;
  amount: number;
}

export type ListTransactionsParams = {
  search?: string;
  limit?: number;
};